% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec2021.R
\name{cec2021}
\alias{cec2021}
\title{CEC 2021}
\usage{
cec2021(i, x, suite)
}
\arguments{
\item{i}{func_number}

\item{x}{input vector}

\item{suite:}{1 - basic
2 - bias
3 - bias-rot
4 - bias-rot-shift
5 - rot
6 - shift
7 - shift-rot
8 - shift-bias}
}
\description{
CEC 2021
}
